+++ plugin/src/main/kotlin/com/github/triplet/gradle/play/internal/Plugins.kt
@@ -21,9 +21,8 @@ internal inline fun <reified T : Task> Project.newTask(
         name: String,
         description: String,
         group: String? = PLUGIN_GROUP,
-        vararg args: Any?,
         block: T.() -> Unit = {}
-): T = tasks.create(name, T::class.java, *args).apply {
+): T = tasks.create(name, T::class.java).apply {
     this.description = description
     this.group = group
     block()